FIXED the problem.
The wrong quotation marks were used when setting the ID of the divs, set the quotations up correctly. The code now displays the donors, in descending order, along with all the information for each donor. The sum of all the donations is now displayed at the top, along with the total amount of donors.